# Optimize buffer sizes for bytestream reads